Made Lock public, moved to sub-package
