--- reaktive/src/commonMain/kotlin/com/badoo/reaktive/utils/ConditionExt.kt
@@ -1,8 +1,0 @@
-package com.badoo.reaktive.utils
-
-internal inline fun <T> Condition.use(block: (Condition) -> T): T =
-    try {
-        block(this)
-    } finally {
-        destroy()
-    }
