+++ reaktive/src/commonTest/kotlin/com/badoo/reaktive/test/Utils.kt
@@ -1,29 +1,10 @@
 package com.badoo.reaktive.test
 
-import com.badoo.reaktive.utils.Condition
 import com.badoo.reaktive.utils.atomic.AtomicReference
-import com.badoo.reaktive.utils.clock.DefaultClock
+import com.badoo.reaktive.utils.lock.Condition
+import com.badoo.reaktive.utils.lock.waitFor
 import com.badoo.reaktive.utils.reaktiveUncaughtErrorHandler
 import kotlin.test.fail
-
-internal fun Condition.waitFor(timeoutNanos: Long, predicate: () -> Boolean): Boolean {
-    require(timeoutNanos > 0L) { "Timeout must be a positive value" }
-
-    val endNanos = DefaultClock.uptimeNanos + timeoutNanos
-    var remainingNanos = timeoutNanos
-
-    while (true) {
-        if (predicate()) {
-            return true
-        }
-        if (remainingNanos <= 0L) {
-            return false
-        }
-
-        await(remainingNanos)
-        remainingNanos = endNanos - DefaultClock.uptimeNanos
-    }
-}
 
 internal fun Condition.waitForOrFail(timeoutNanos: Long, predicate: () -> Boolean) {
     if (!waitFor(timeoutNanos, predicate)) {
