+++ reaktive/src/jsMain/kotlin/com/badoo/reaktive/utils/lock/Lock.kt
@@ -1,0 +1,20 @@
+package com.badoo.reaktive.utils.lock
+
+actual class Lock {
+
+    actual fun acquire() {
+        // no-op
+    }
+
+    actual fun release() {
+        // no-op
+    }
+
+    actual fun destroy() {
+        // no-op
+    }
+
+    actual fun newCondition(): Condition {
+        throw RuntimeException("Condition is not supported in JavaScript")
+    }
+}
