+++ reaktive/src/jvmNativeCommonTest/kotlin/com/badoo/reaktive/test/DoInBackground.kt
@@ -1,8 +1,8 @@
 package com.badoo.reaktive.test
 
-import com.badoo.reaktive.utils.Lock
+import com.badoo.reaktive.utils.lock.Lock
 import com.badoo.reaktive.utils.atomic.AtomicBoolean
-import com.badoo.reaktive.utils.synchronized
+import com.badoo.reaktive.utils.lock.synchronized
 
 internal expect fun doInBackground(block: () -> Unit)
